Update varlist.json
All sechiba.card input files are not processed into compressed format for ORCHIDEE-CNP. If not done, the MLA produces global restart files from compressed training data which are far from the equilibrium. This is likely due to issues in ORCHIDEE with the writing or reading of restart files (information is not read from restart file but from input files).

The processing of compressed files from the global ones is fast, thus this provides little detoriation in the comp. effieciency.